Complete rewrite of regular
Regular was redesigned and simplified so that counter and fixed can be
reduced to it. It now resembles a labelled transition system.

It was necessary to separate the instructions from State.t and to re-
design Graph.t. Furthermore the field State.Ok.count was removed be-
cause it can be derived from the other invariants.