Initial draft of TransitionSystem.v
We use Typeclasses because you can not express that the image of a
homomorphism is a model of a theory with the module system.

After much hesitation I decided to parametrize signatures over Types
instead of Setoids or PartialSetoids and use a separate Typeclass for
equalities. This approach is much smoother but could lead to ambiguity
if one adds inappropriate instances to the context.